+++ project-model/src/main/java/com/walmartlabs/concord/project/yaml/model/YamlProfile.java
@@ -20,10 +20,14 @@ package com.walmartlabs.concord.project.yaml.model;
  * =====
  */
 
-import com.fasterxml.jackson.annotation.*;
+import com.fasterxml.jackson.annotation.JsonCreator;
+import com.fasterxml.jackson.annotation.JsonProperty;
 
 import java.io.Serializable;
-import java.util.*;
+import java.util.HashMap;
+import java.util.List;
+import java.util.Map;
+import java.util.Objects;
 import java.util.stream.Collectors;
 
 public class YamlProfile implements Serializable {